separate bdv functionality
