+++ src/main/java/de/embl/cba/elastixwrapper/wrapper/BdvManager.java
@@ -1,0 +1,83 @@
+package de.embl.cba.elastixwrapper.wrapper;
+
+import bdv.util.*;
+import ij.ImagePlus;
+import ij.measure.Calibration;
+import net.imglib2.RandomAccessibleInterval;
+import net.imglib2.img.display.imagej.ImageJFunctions;
+import net.imglib2.type.numeric.ARGBType;
+
+import java.io.File;
+import java.util.ArrayList;
+
+public class BdvManager {
+
+    private ArrayList<ARGBType> colors;
+    private Bdv bdv;
+    private int colorIndex;
+
+    public BdvManager() {
+        initColors();
+    }
+
+    private void initColors()
+    {
+        colors = new ArrayList<>();
+        colors.add( new ARGBType( ARGBType.rgba( 000, 255, 000, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 000, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 000, 000, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 000, 000, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 000, 255, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 255, 000, 255 ) ) );
+
+        colors.add( new ARGBType( ARGBType.rgba( 255, 255, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 255, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 255, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 255, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 255, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 255, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 255, 255, 255 ) ) );
+        colors.add( new ARGBType( ARGBType.rgba( 255, 255, 255, 255 ) ) );
+        colorIndex = 0;
+    }
+
+    public Bdv showMetaImageInBdv( String imageDir, String imageName )  {
+        ImagePlus imagePlus = loadMetaImage( imageDir, imageName );
+        final BdvStackSource bdvStackSource = showImagePlusInBdv( imagePlus );
+        bdvStackSource.setColor( colors.get( colorIndex++ )  );
+        bdv = bdvStackSource.getBdvHandle();
+        return bdv;
+    }
+
+    public BdvStackSource showImagePlusInBdv(
+            ImagePlus imp )
+    {
+        final Calibration calibration = imp.getCalibration();
+
+        if ( imp.getNSlices() > 1 )
+        {
+            final double[] calib = {
+                    calibration.pixelWidth,
+                    calibration.pixelHeight,
+                    calibration.pixelDepth
+            };
+            return BdvFunctions.show(
+                    (RandomAccessibleInterval) ImageJFunctions.wrapReal( imp ),
+                    imp.getTitle(),
+                    BdvOptions.options().addTo( bdv ).axisOrder( AxisOrder.XYZ ).sourceTransform( calib ) );
+        }
+        else
+        {
+
+            final double[] calib = {
+                    calibration.pixelWidth,
+                    calibration.pixelHeight
+            };
+
+            return BdvFunctions.show(
+                    ( RandomAccessibleInterval ) ImageJFunctions.wrapReal( imp ),
+                    imp.getTitle(),
+                    BdvOptions.options().addTo( bdv ).is2D().sourceTransform( calib ) );
+        }
+    }
+}
